Changes to Spacemove
I'm trying to make gravity scale properly. I experimented with a
different equation, but commented it out and went back to realistic
gravity. My current tactic is going to be making the main thruster weak
and have low fuel, but give the player a "boost" button that blasts them
forward very quickly, but can only be used near planets and has a
several second cooldown. Next task is a cooldown for the regular
thruster. Then multiplayer. More long term, I'll need UI, weapons,
flags, scoring, and a match system. After that, the game will be fully
playable, and I can start adding "fluffy" things like interesting levels
and any graphics at all.

+++ Assets/Scripts/Gravity/Gravity.cs
@@ -19,10 +19,12 @@ public class Gravity : MonoBehaviour
         //Iterate through gravityTransforms (and gravityMasses) to apply the forces
         foreach(GravityMass gravObj in GravityObjects.gravList)
         {
+            //Debug.Log(gravObj.name + " " + Vector3.Distance(transform.position, gravObj.transform.position));
             rigbod.AddForce(GetGravityForce(gravObj.GetPos(), gravObj.GetMass()), ForceMode.Impulse);
         }
     }
 
+    //Method to calculate force of gravity
     private Vector3 GetGravityForce(Vector3 pos, float mass)
     {
         //Get the direction to the attractor
@@ -37,7 +39,25 @@ public class Gravity : MonoBehaviour
         //Force is the direction * the strength of the pull
         //Because we're ignoring the Gravitational Constant, 
         //pull strength = our mass * other mass / distance squared
-        Vector3 force = dir * (float)((rigbod.mass * mass) / distSqrd);
+        Vector3 force = dir * ((rigbod.mass * mass) / distSqrd);
+        //Debug.Log((rigbod.mass * mass) / distSqrd);
+        
+        //Deprecated--more complex distance term
+        //Force is direction * the strength of the pull
+        //Because we're ignoring the gravitational constant, 
+        //pull strength = our mass * other mass * distance modifier
+        //Vector3 force = dir * ((rigbod.mass * mass) * CalcDistanceTerm(Vector3.Distance(transform.position, pos), radius));
         return force;
     }
+
+    //Deprecated
+    //This is a complex method to calculate the distance term for gravity so it scales exactly how I want it to
+    //It's a lot more complicated than 1/x^2
+    //Essentially, I want gravity to be fairly strong around the planet, but start dropping off slowly farther away, then go to near zero
+    //It looks like this: ((-arctan((x - 40) / 10)) / (pi / 2)) + 1) / 2
+    //private float CalcDistanceTerm(float dist, float radius)
+    //{
+    //    Debug.Log(((-Mathf.Atan(dist - radius) / (radius / 4)) / (Mathf.PI / 2) + 1) / 2);
+    //    return ((-Mathf.Atan(dist - radius) / (radius / 4)) / (Mathf.PI / 2) + 1) / 2;
+    //}
 }
